+++ packages/templates/cra-template-desktop-viewer/template/src/frontend/index.tsx
@@ -23,7 +23,6 @@ const viewerFrontendMain = async () => {
   // when initialization is complete, render
   const container = document.getElementById("root") as HTMLElement;
   const root = createRoot(container);
-  UiFramework.childWindows.useCreateRoot(createRoot);
 
   document.documentElement.classList.add(`iui-theme-dark`);